cloudfunctions_requirements.txt
google-cloud-storage==2.14.0
google-cloud-logging==3.9.0
functions-framework==3.5.0
pandas==2.2.0
python-dotenv==1.0.0
requests==2.31.0
google-genai>=1.0.0
googlemaps==4.10.0
fsspec==2023.12.2
pyarrow==15.0.0
rapidfuzz==3.6.1
google-cloud-tasks==2.16.0
